pass test_csv: Support QUOTE_NONNUMERIC
Changes `escape` to return payload unchanged if it is a number (or
boolean). This is to pass the QUOTE_NONNUMERIC test[0] in test_csv.

Also adds return statements to `_ProxyWriter.writerow` and
`_ProxyWriter.writerows` to pass the tests.

Added a new test function `test_safe_nonstr_sample_payloads` to handle
those cases.

[0]: https://github.com/python/cpython/blob/89192c46da7b984811ff3bd648f8e827e4ef053c/Lib/test/test_csv.py#L182-L183